+++ Client/CHANGELOG.md
@@ -16,4 +16,8 @@ These fixes generally resolve issues that don't affect the software's overall fu
 
 ## 0.0.1
 - set up ItemView and App.vue for demo purposes
-- set up dark mode toggle
+- set up dark mode toggle 
+
+## 1.0.0
+- set up frontend to handle bulk inserts
+- moved all api logic to itemStore to manage state